Tests: Add library unit tests

+++ src/main/java/org/semux/util/IOUtil.java
@@ -24,7 +24,7 @@ import java.util.List;
 public class IOUtil {
 
     /**
-     * Read stream into byte array, and close it afterwards.
+     * Reads stream into byte array, and close it afterwards.
      * 
      * @param in
      * @return
@@ -43,7 +43,7 @@ public class IOUtil {
     }
 
     /**
-     * Read stream into a string, and close it afterwards.
+     * Reads stream a string, and close it afterwards.
      * 
      * @param in
      * @return
@@ -54,29 +54,35 @@ public class IOUtil {
     }
 
     /**
-     * Read file as a byte array.
+     * Reads file as a byte array.
      * 
      * @param file
-     * @return
+     *            The file to read
+     * @return a byte array; empty array if the file does not exist
      * @throws IOException
      */
     public static byte[] readFile(File file) throws IOException {
-        InputStream in = new BufferedInputStream(new FileInputStream(file));
-        ByteArrayOutputStream out = new ByteArrayOutputStream();
+        if (file.exists()) {
+            InputStream in = new BufferedInputStream(new FileInputStream(file));
+            ByteArrayOutputStream out = new ByteArrayOutputStream();
 
-        for (int c; (c = in.read()) != -1;) {
-            out.write(c);
+            for (int c; (c = in.read()) != -1;) {
+                out.write(c);
+            }
+            in.close();
+
+            return out.toByteArray();
+        } else {
+            return Bytes.EMPTY_BYTES;
         }
-        in.close();
-
-        return out.toByteArray();
     }
 
     /**
-     * Read file as a string.
+     * Reads file as a string.
      * 
      * @param file
-     * @return
+     *            File to read
+     * @return The file content as string; empty if the file does not exist
      * @throws IOException
      */
     public static String readFileAsString(File file) throws IOException {
@@ -84,10 +90,12 @@ public class IOUtil {
     }
 
     /**
-     * Write a byte array into a file.
+     * Writes a byte array into a file.
      * 
      * @param bytes
+     *            A byte array
      * @param file
+     *            Destination file
      * @throws IOException
      */
     public static void writeToFile(byte[] bytes, File file) throws IOException {
@@ -97,10 +105,12 @@ public class IOUtil {
     }
 
     /**
-     * Write a stirng into a file.
+     * Writes a {@link String} into a file.
      * 
      * @param str
+     *            String contents
      * @param file
+     *            Destination file
      * @throws IOException
      */
     public static void writeToFile(String str, File file) throws IOException {
@@ -108,32 +118,41 @@ public class IOUtil {
     }
 
     /**
-     * Read file line by line.
+     * Reads file by lines.
      * 
      * @param file
-     * @return
+     *            The file to read
+     * @return A list of lines in the file, or empty if the file doesn't exist
      * @throws IOException
      */
     public static List<String> readFileAsLines(File file) throws IOException {
         List<String> lines = new ArrayList<>();
 
-        try (BufferedReader in = new BufferedReader(new FileReader(file))) {
-            for (String line; (line = in.readLine()) != null;) {
-                lines.add(line);
+        if (file.isFile()) {
+            try (BufferedReader in = new BufferedReader(new FileReader(file))) {
+                for (String line; (line = in.readLine()) != null;) {
+                    lines.add(line);
+                }
             }
-            return lines;
         }
+
+        return lines;
     }
 
     /**
      * Copy a file to a new location.
      * 
      * @param src
+     *            Source file
      * @param dst
-     * @param override
+     *            Destination file
+     * @param replaceExisting
+     *            Whether to replace existing file
      * @throws IOException
      */
-    public static void copyFile(File src, File dst, boolean override) throws IOException {
-        Files.copy(src.toPath(), dst.toPath(), StandardCopyOption.REPLACE_EXISTING);
+    public static void copyFile(File src, File dst, boolean replaceExisting) throws IOException {
+        if (replaceExisting || !dst.exists()) {
+            Files.copy(src.toPath(), dst.toPath(), StandardCopyOption.REPLACE_EXISTING);
+        }
     }
 }
